In commit-pin-updates, do 'ci' as a new make invocation
With some repos, including confd, there is a 'clean' hiding inside the
ci target, which can remove some of the dependencies needed for
testing.  In the confd case, 'clean' removes bin/confd.

The testing targets after the 'clean' have correct dependencies, but
the problem is that make doesn't recheck that they still exist, if it
knows that they were built earlier during the same make invocation.
So bin/confd (for example) does not get built again after the 'clean',
and the confd test then fails with

    /tests/test_suite_common.sh: line 377: confd: command not found

To fix this, rewrite so that the 'ci' is a fresh make invocation.