Fix s390x dynamic loader symlink
This change resolves an issue with the s390x dynamic loader symlink.
On s390x UBI 8, the symlink points to `/usr/lib64`. This directory
is not included in the Calico base image, resulting a dangling symlink
and preventing programs from loading. This update corrects the
symlink and added a test to validate the minimum C runtime. It also
fixes semvalidator multi-arch builds.